Split `SideColumnNavBar` into two separate components
